# Perf: optimize VerifyPackage by not querying WM feed unless necessary