Issue #5049 - optimize upload submit by not querying `CuratedFeedService.GetFeedByName` unless necessary. Profile showed 46% inclusive in submit due to this query.

Scenario: upload of package that won't be curated
- Before: 11.52s, 10.61s, 12.71s => 11.61s
- After: 2.74s, 2.36s, 2.31s => 2.47s (-78.73%)